+++ persistence/zeromq.go
@@ -90,7 +90,7 @@ func (instance *zeromq) GetNumberOfTorrents() (uint, error) {
 	return 0, nil
 }
 
-func (instance *zeromq) GetNumberOfQueryTorrents(query string, epoch int64) (uint, error) {
+func (instance *zeromq) GetNumberOfQueryTorrents(query string, epoch int64) (uint64, error) {
 	return 0, nil
 }
